[#263] cache control 로직을 http mod로 분리
