+++ rupring/src/http/mod.rs
@@ -1,1 +1,2 @@
+pub mod cache;
 pub mod meme;
